bot_data: Validate parameters with Zod.
