+++ web/src/bot_data.ts
@@ -1,6 +1,8 @@
-import {z} from "zod";
+import type {z} from "zod";
 
+import {server_add_bot_schema, server_update_bot_schema, services_schema} from "./bot_types";
 import * as people from "./people";
+import type {StateData} from "./state_data";
 
 export type ServerUpdateBotData = z.infer<typeof server_update_bot_schema>;
 export type ServerAddBotData = z.infer<typeof server_add_bot_schema>;
@@ -8,59 +10,16 @@ export type Bot = Omit<ServerAddBotData, "services">;
 
 export type Services = z.infer<typeof services_schema>;
 
-export type BotDataParams = {
-    realm_bots: ServerAddBotData[];
-};
-
 const bots = new Map<number, Bot>();
 const services = new Map<number, Services>();
-
-// Define zod schema for data validation
-const basic_bot_schema = z.object({
-    api_key: z.string(),
-    avatar_url: z.string(),
-    bot_type: z.number(),
-    default_all_public_streams: z.boolean(),
-    default_events_register_stream: z.string().nullable(),
-    default_sending_stream: z.string().nullable(),
-    email: z.string(),
-    full_name: z.string(),
-    is_active: z.boolean(),
-    owner_id: z.number().nullable(),
-    user_id: z.number(),
-});
-
-const outgoing_service_schema = z.object({
-    base_url: z.string(),
-    interface: z.number(),
-    token: z.string(),
-});
-
-const embedded_service_schema = z.object({
-    config_data: z.record(z.string()),
-    service_name: z.string(),
-});
-
-const services_schema = z.union([
-    z.array(outgoing_service_schema),
-    z.array(embedded_service_schema),
-]);
-
-const server_update_bot_schema = basic_bot_schema.extend({
-    services: services_schema,
-});
-
-const server_add_bot_schema = server_update_bot_schema.extend({
-    bot_type: z.number(),
-    email: z.string(),
-    is_active: z.boolean(),
-});
 
 export function all_user_ids(): number[] {
     return [...bots.keys()];
 }
 
 export function add(bot_data: ServerAddBotData): void {
+    // TODO/typescript: Move validation to the caller when
+    // server_events_dispatch.js is converted to TypeScript.
     const {services: bot_services, ...clean_bot} = server_add_bot_schema.parse(bot_data);
     bots.set(clean_bot.user_id, clean_bot);
 
@@ -74,6 +33,8 @@ export function del(bot_id: number): void {
 
 export function update(bot_id: number, bot_update: ServerUpdateBotData): void {
     const bot = bots.get(bot_id)!;
+    // TODO/typescript: Move validation to the caller when
+    // server_events_dispatch.js is converted to TypeScript.
     Object.assign(bot, server_update_bot_schema.deepPartial().parse(bot_update));
 
     // We currently only support one service per bot.
@@ -125,7 +86,7 @@ export function get_services(bot_id: number): Services | undefined {
     return services.get(bot_id);
 }
 
-export function initialize(params: BotDataParams): void {
+export function initialize(params: StateData["bot"]): void {
     bots.clear();
     for (const bot of params.realm_bots) {
         add(bot);
